requirements.txt
os
tqdm
google-generativeai
argparse
json
shutil
re
pillow